Split test_views into the various files corresponding to the codebase structure.
Also created a base class for data and a factory for the null tolerance tests.

Files checked in: DataRepo/tests/test_views.py DataRepo/tests/views/models/base.py DataRepo/tests/views/models/test_animal.py DataRepo/tests/views/models/test_archive_file.py DataRepo/tests/views/models/test_compound.py DataRepo/tests/views/models/test_infusate.py DataRepo/tests/views/models/test_msrun_sample.py DataRepo/tests/views/models/test_msrun_sequence.py DataRepo/tests/views/models/test_peak_data.py DataRepo/tests/views/models/test_peak_group.py DataRepo/tests/views/models/test_sample.py DataRepo/tests/views/models/test_study.py DataRepo/tests/views/models/test_tissue.py DataRepo/tests/views/search/test_advanced.py
