+++ DataRepo/tests/views/models/test_compound.py
@@ -1,0 +1,37 @@
+from django.test import tag
+from django.urls import reverse
+
+from DataRepo.models import Compound
+from DataRepo.tests.views.models.base import (
+    ModelViewTests,
+    create_null_tolerance_test_class,
+)
+
+
+class CompoundViewTests(ModelViewTests):
+    @tag("compound")
+    def test_compound_list(self):
+        response = self.client.get(reverse("compound_list"))
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/compound_list.html")
+        self.assertEqual(
+            len(response.context["compound_list"]), self.ALL_COMPOUNDS_COUNT
+        )
+
+    @tag("compound")
+    def test_compound_detail(self):
+        lysine = Compound.objects.filter(name="lysine").get()
+        response = self.client.get(reverse("compound_detail", args=[lysine.id]))
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/compound_detail.html")
+        self.assertEqual(response.context["compound"].name, "lysine")
+
+    @tag("compound")
+    def test_compound_detail_404(self):
+        c = Compound.objects.order_by("id").last()
+        response = self.client.get(reverse("compound_detail", args=[c.id + 1]))
+        self.assertEqual(response.status_code, 404)
+
+
+# This runs the above tests again with auto-update diabled
+CompoundViewNullToleranceTests = create_null_tolerance_test_class(CompoundViewTests)
