+++ DataRepo/tests/views/models/test_peak_data.py
@@ -1,0 +1,30 @@
+from DataRepo.models import PeakData, PeakGroup
+from DataRepo.tests.views.models.base import (
+    ModelViewTests,
+    create_null_tolerance_test_class,
+)
+
+
+class PeakDataViewTests(ModelViewTests):
+    def test_peakdata_list(self):
+        """
+        the total rows loaded may be greater than total rows in file for each sample,
+        since rows are created during loading for missing labeled_count
+        """
+        response = self.client.get("/DataRepo/peakdata/")
+        pd = PeakData.objects.all()
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/peakdata_list.html")
+        self.assertEqual(len(response.context["peakdata_list"]), pd.count())
+
+    def test_peakdata_list_per_peakgroup(self):
+        pg1 = PeakGroup.objects.filter(msrun_sample__sample__name="serum-xz971").first()
+        pd1 = PeakData.objects.filter(peak_group_id=pg1.pk)
+        response = self.client.get("/DataRepo/peakdata/?peak_group_id=" + str(pg1.pk))
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/peakdata_list.html")
+        self.assertEqual(len(response.context["peakdata_list"]), pd1.count())
+
+
+# This runs the above tests again with auto-update diabled
+PeakDataViewNullToleranceTests = create_null_tolerance_test_class(PeakDataViewTests)
