+++ DataRepo/tests/views/models/test_tissue.py
@@ -1,0 +1,35 @@
+from django.test import tag
+from django.urls import reverse
+
+from DataRepo.models import Tissue
+from DataRepo.tests.views.models.base import (
+    ModelViewTests,
+    create_null_tolerance_test_class,
+)
+
+
+class TissueViewTests(ModelViewTests):
+    @tag("tissue")
+    def test_tissue_list(self):
+        response = self.client.get(reverse("tissue_list"))
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/tissue_list.html")
+        self.assertEqual(len(response.context["tissue_list"]), self.ALL_TISSUES_COUNT)
+
+    @tag("tissue")
+    def test_tissue_detail(self):
+        t1 = Tissue.objects.filter(name="brown_adipose_tissue").get()
+        response = self.client.get(reverse("tissue_detail", args=[t1.id]))
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/tissue_detail.html")
+        self.assertEqual(response.context["tissue"].name, "brown_adipose_tissue")
+
+    @tag("tissue")
+    def test_tissue_detail_404(self):
+        t = Tissue.objects.order_by("id").last()
+        response = self.client.get(reverse("tissue_detail", args=[t.id + 1]))
+        self.assertEqual(response.status_code, 404)
+
+
+# This runs the above tests again with auto-update diabled
+TissueViewNullToleranceTests = create_null_tolerance_test_class(TissueViewTests)
