+++ DataRepo/tests/views/search/test_advanced.py
@@ -1,0 +1,178 @@
+import json
+
+from DataRepo.models import PeakGroup
+from DataRepo.tests.views.models.base import (
+    ModelViewTests,
+    create_null_tolerance_test_class,
+)
+
+
+class AdvancedSearchViewTests(ModelViewTests):
+    def test_search_advanced_browse(self):
+        """
+        Load the advanced search page in browse mode and make sure the mode is added to the context data
+        """
+        response = self.client.get(
+            "/DataRepo/search_advanced/?mode=browse&format=pdtemplate"
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/search/query.html")
+        self.assertEqual(response.context["mode"], "browse")
+        self.assertEqual(response.context["format"], "pdtemplate")
+
+    def test_search_advanced_search(self):
+        """
+        Load the advanced search page in the default search mode and make sure the mode is added to the context data
+        """
+        response = self.client.get("/DataRepo/search_advanced/")
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/search/query.html")
+        self.assertEqual(response.context["mode"], "search")
+
+    def get_advanced_search_inputs(self):
+        asform = {
+            "fmt": "pgtemplate",
+            "form-TOTAL_FORMS": "3",
+            "form-INITIAL_FORMS": "0",
+            "form-0-pos": "pgtemplate-PeakGroups-selected.0-all-False.0",
+            "form-0-fld": "msrun_sample__sample__tissue__name",
+            "form-0-ncmp": "iexact",
+            "form-0-val": "Brain",
+            "form-0-units": "identity",
+            "form-1-pos": "pdtemplate-PeakData.0-all-False.0",
+            "form-1-fld": "labels__element",
+            "form-1-ncmp": "iexact",
+            "form-1-units": "identity",
+            "form-2-pos": "fctemplate-FCirc.0-all-False.0",
+            "form-2-fld": "msrun_sample__sample__animal__name",
+            "form-2-ncmp": "iexact",
+            "form-2-units": "identity",
+        }
+        qry = self.get_advanced_qry()
+        dlform = {
+            "form-TOTAL_FORMS": "1",
+            "form-INITIAL_FORMS": "0",
+            "qryjson": json.dumps(qry),
+        }
+        return [asform, qry, dlform]
+
+    def get_advanced_qry(self):
+        """
+        Create a simple advanced query
+        """
+        return {
+            "selectedtemplate": "pgtemplate",
+            "searches": {
+                "pgtemplate": {
+                    "tree": {
+                        "pos": "",
+                        "type": "group",
+                        "val": "all",
+                        "static": False,
+                        "queryGroup": [
+                            {
+                                "type": "query",
+                                "pos": "",
+                                "fld": "msrun_sample__sample__tissue__name",
+                                "ncmp": "iexact",
+                                "static": "",
+                                "val": "Brain",
+                                "units": "identity",
+                            }
+                        ],
+                    },
+                    "name": "PeakGroups",
+                },
+                "pdtemplate": {
+                    "tree": {
+                        "pos": "",
+                        "type": "group",
+                        "val": "all",
+                        "static": False,
+                        "queryGroup": [
+                            {
+                                "type": "query",
+                                "pos": "",
+                                "ncmp": "iexact",
+                                "static": "",
+                                "fld": "labels__element",
+                                "val": "",
+                                "units": "identity",
+                            }
+                        ],
+                    },
+                    "name": "PeakData",
+                },
+                "fctemplate": {
+                    "tree": {
+                        "pos": "",
+                        "type": "group",
+                        "val": "all",
+                        "static": False,
+                        "queryGroup": [
+                            {
+                                "type": "query",
+                                "pos": "",
+                                "fld": "msrun_sample__sample__animal__name",
+                                "ncmp": "iexact",
+                                "static": "",
+                                "val": "",
+                                "units": "identity",
+                            }
+                        ],
+                    },
+                    "name": "FCirc",
+                },
+            },
+        }
+
+    def test_search_advanced_valid(self):
+        """
+        Do a simple advanced search and make sure the results are correct
+        """
+        qs = PeakGroup.objects.filter(
+            msrun_sample__sample__tissue__name__iexact="Brain"
+        ).prefetch_related("msrun_sample__sample__animal__studies")
+        [filledform, qry, ignore] = self.get_advanced_search_inputs()
+        response = self.client.post("/DataRepo/search_advanced/", filledform)
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/search/query.html")
+        self.assertEqual(len(response.context["res"]), qs.count())
+        self.assertEqual(qry, response.context["qry"])
+
+    def test_search_advanced_invalid(self):
+        """
+        Do a simple advanced search and make sure the results are correct
+        """
+        [invalidform, qry, ignore] = self.get_advanced_search_inputs()
+        # Make the form invalid
+        invalidform.pop("form-0-val", None)
+        # Expected response difference:
+        qry["searches"]["pgtemplate"]["tree"]["queryGroup"][0]["val"] = ""
+        response = self.client.post("/DataRepo/search_advanced/", invalidform)
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed(response, "DataRepo/search/query.html")
+        self.assertEqual(len(response.context["res"]), 0)
+        self.assertEqual(qry, response.context["qry"])
+
+    def test_search_advanced_tsv(self):
+        """
+        Download a simple advanced search and make sure the results are correct
+        """
+        [_, _, dlform] = self.get_advanced_search_inputs()
+        response = self.client.post("/DataRepo/search_advanced_tsv/", dlform)
+
+        # Response content settings
+        self.assertEqual(response.get("Content-Type"), "application/text")
+        self.assertEqual(response.status_code, 200)
+        # Cannot use assertContains here for non-http response - it will complain about a missing status_code
+        contentdisp = response.get("Content-Disposition")
+        self.assertTrue("attachment" in contentdisp)
+        self.assertTrue("PeakGroups" in contentdisp)
+        self.assertTrue(".tsv" in contentdisp)
+
+
+# This runs the above tests again with auto-update diabled
+AdvancedSearchViewNullToleranceTests = create_null_tolerance_test_class(
+    AdvancedSearchViewTests
+)
